+++ OUDS/Core/OUDS/Tests/OUDSTheme/MockThemes/MockTheme+SpaceSemanticTokens.swift
@@ -72,98 +72,98 @@ extension MockTheme {
 
     // MARK: Semantic token - Spacing - Padding - Padding inline
 
-    override open var spacePaddingInlineNone: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineShorter: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineShort: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineMedium: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineTall: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineTaller: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineTallest: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineNone: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineShorter: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineShort: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineMedium: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineTall: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineTaller: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineTallest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
 
-    override open var spacePaddingInlineWithIconNone: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithIconShortest: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithIconShorter: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithIconShort: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithIconMedium: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithIconTall: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithIconTaller: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithIconTallest: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithIconNone: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithIconShortest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithIconShorter: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithIconShort: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithIconMedium: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithIconTall: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithIconTaller: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithIconTallest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
 
-    override open var spacePaddingInlineWithArrowNone: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithArrowShortest: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithArrowShorter: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithArrowShort: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithArrowMedium: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithArrowTall: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithArrowTaller: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingInlineWithArrowTallest: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithArrowNone: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithArrowShortest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithArrowShorter: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithArrowShort: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithArrowMedium: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithArrowTall: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithArrowTaller: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingInlineWithArrowTallest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
 
     // MARK: Semantic token - Padding - Padding stack
 
-    override open var spacePaddingBlockNone: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockShorter: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockShort: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockMedium: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockTall: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockTaller: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockTallest: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockNone: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockShorter: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockShort: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockMedium: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockTall: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockTaller: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockTallest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
 
-    override open var spacePaddingBlockWithIconNone: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockWithIconShortest: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockWithIconShorter: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockWithIconShort: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockWithIconMedium: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockWithIconTall: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spacePaddingBlockWithIconTaller: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockWithIconNone: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockWithIconShortest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockWithIconShorter: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockWithIconShort: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockWithIconMedium: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockWithIconTall: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spacePaddingBlockWithIconTaller: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
 
     // MARK: Semantic token - Padding - Padding inset
 
-    override open var spaceInsetNone: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceInsetSmash: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceInsetShortest: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceInsetShorter: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceInsetShort: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceInsetMedium: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceInsetTall: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceInsetTaller: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceInsetTallest: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceInsetSpacious: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceInsetNone: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceInsetSmash: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceInsetShortest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceInsetShorter: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceInsetShort: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceInsetMedium: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceInsetTall: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceInsetTaller: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceInsetTallest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceInsetSpacious: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
 
     // MARK: Semantic token - Padding - Gap inline
 
-    override open var spaceColumnGapNone: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapShorter: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapShort: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapMedium: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapTall: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapTaller: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapNone: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapShorter: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapShort: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapMedium: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapTall: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapTaller: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
 
-    override open var spaceColumnGapWithIconNone: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapWithIconShortest: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapWithIconShorter: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapWithIconShort: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapWithIconMedium: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapWithIconTall: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithIconNone: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithIconShortest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithIconShorter: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithIconShort: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithIconMedium: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithIconTall: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
 
-    override open var spaceColumnGapWithArrowNone: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapWithArrowShortest: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapWithArrowShorter: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapWithArrowShort: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapWithArrowMedium: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceColumnGapWithArrowTall: SpacePaddingInlineSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithArrowNone: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithArrowShortest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithArrowShorter: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithArrowShort: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithArrowMedium: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceColumnGapWithArrowTall: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
 
     // MARK: Semantic token - Padding - Gap stack
 
-    override open var spaceRowGapNone: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceRowGapShortest: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceRowGapShorter: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceRowGapShort: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceRowGapMedium: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceRowGapTall: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceRowGapWithIconNone: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceRowGapWithIconShortest: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceRowGapWithIconShorter: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceRowGapWithIconShort: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceRowGapWithIconMedium: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
-    override open var spaceRowGapWithIconTall: SpaceGapStackSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapNone: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapShortest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapShorter: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapShort: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapMedium: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapTall: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapWithIconNone: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapWithIconShortest: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapWithIconShorter: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapWithIconShort: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapWithIconMedium: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
+    override open var spaceRowGapWithIconTall: SpaceSemanticToken { Self.mockThemeSpaceRawToken }
 }